Move `emit_mir` out of `rustc_mir_transform::dump_mir`
This function doesn't have an obvious home, but there's little reason for it to
be in `rustc_mir_transform`, and having it in `rustc_driver_impl::pretty` at
least puts it near other callers of `write_mir_pretty`.
